move endpoints list to their own object
