+++ src/main/scala/com/softwaremill/realworld/Main.scala
@@ -23,7 +23,7 @@ object Main extends ZIOAppDefault:
           )
         )
         .options
-    val app: HttpApp[Any, Throwable] = ZioHttpInterpreter(serverOptions).toHttp(endpoints)
+    val app: HttpApp[Any, Throwable] = ZioHttpInterpreter(serverOptions).toHttp(Endpoints.endpoints)
 
     val port = sys.env.get("HTTP_PORT").flatMap(_.toIntOption).getOrElse(8080)
